+++ projection/winrt-Microsoft.UI.Xaml.Interop/winrt/microsoft/ui/xaml/interop/__init__.py
@@ -34,5 +34,5 @@ IBindableObservableVector = _winrt_microsoft_ui_xaml_interop.IBindableObservable
 IBindableVector = _winrt_microsoft_ui_xaml_interop.IBindableVector
 IBindableVectorView = _winrt_microsoft_ui_xaml_interop.IBindableVectorView
 INotifyCollectionChanged = _winrt_microsoft_ui_xaml_interop.INotifyCollectionChanged
-BindableVectorChangedEventHandler = typing.Callable[[typing.Optional[IBindableObservableVector], typing.Optional[winrt.system.Object]], None]
-NotifyCollectionChangedEventHandler = typing.Callable[[typing.Optional[winrt.system.Object], typing.Optional[NotifyCollectionChangedEventArgs]], None]
+BindableVectorChangedEventHandler = typing.Callable[[IBindableObservableVector, winrt.system.Object], None]
+NotifyCollectionChangedEventHandler = typing.Callable[[winrt.system.Object, NotifyCollectionChangedEventArgs], None]