+++ projection/winrt-Windows.Devices.Perception.Provider/winrt/windows/devices/perception/provider/__init__.py
@@ -35,5 +35,5 @@ PerceptionPropertyChangeRequest = _winrt_windows_devices_perception_provider.Per
 PerceptionVideoFrameAllocator = _winrt_windows_devices_perception_provider.PerceptionVideoFrameAllocator
 IPerceptionFrameProvider = _winrt_windows_devices_perception_provider.IPerceptionFrameProvider
 IPerceptionFrameProviderManager = _winrt_windows_devices_perception_provider.IPerceptionFrameProviderManager
-PerceptionStartFaceAuthenticationHandler = typing.Callable[[typing.Optional[PerceptionFaceAuthenticationGroup]], bool]
-PerceptionStopFaceAuthenticationHandler = typing.Callable[[typing.Optional[PerceptionFaceAuthenticationGroup]], None]
+PerceptionStartFaceAuthenticationHandler = typing.Callable[[PerceptionFaceAuthenticationGroup], bool]
+PerceptionStopFaceAuthenticationHandler = typing.Callable[[PerceptionFaceAuthenticationGroup], None]