+++ projection/winrt-Windows.UI.Xaml.Media.Imaging/winrt/windows/ui/xaml/media/imaging/__init__.py
@@ -52,4 +52,4 @@ SvgImageSourceOpenedEventArgs = _winrt_windows_ui_xaml_media_imaging.SvgImageSou
 VirtualSurfaceImageSource = _winrt_windows_ui_xaml_media_imaging.VirtualSurfaceImageSource
 WriteableBitmap = _winrt_windows_ui_xaml_media_imaging.WriteableBitmap
 XamlRenderingBackgroundTask = _winrt_windows_ui_xaml_media_imaging.XamlRenderingBackgroundTask
-DownloadProgressEventHandler = typing.Callable[[typing.Optional[winrt.system.Object], typing.Optional[DownloadProgressEventArgs]], None]
+DownloadProgressEventHandler = typing.Callable[[winrt.system.Object, DownloadProgressEventArgs], None]